# backend/openapi.yaml
openapi: 3.0.0
paths:
  /shanda/auth/webhook:
    post:
      operationId: ClerkAuthController_handleClerkWebhook
      parameters:
        - name: svix-id
          required: true
          in: header
          schema:
            type: string
        - name: svix-timestamp
          required: true
          in: header
          schema:
            type: string
        - name: svix-signature
          required: true
          in: header
          schema:
            type: string
      responses:
        '201':
          description: ''
      summary: Clerk webhook handler
      tags: &ref_0
        - auth
  /shanda/auth/me:
    get:
      operationId: ClerkAuthController_getCurrentUser
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      security:
        - bearer: []
      summary: Get current user profile
      tags: *ref_0
  /shanda/auth/profile:
    put:
      operationId: ClerkAuthController_updateProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      security:
        - bearer: []
      summary: Update user profile
      tags: *ref_0
  /shanda/auth/roles:
    put:
      operationId: ClerkAuthController_updateRoles
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      security:
        - bearer: []
      summary: Update user roles
      tags: *ref_0
...
  /shanda/jobs:
    get:
      operationId: JobsController_findAll
      parameters:
        - name: page
          required: false
          in: query
          description: 'Page number (default: 1)'
          schema:
            minimum: 1
            default: 1
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: 'Items per page (default: 20, max: 100)'
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: number
      responses:
        '200':
          description: Paginated list of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobResponseDto'
      security: &ref_8
        - bearer: []
      summary: Get all jobs (owners see theirs, providers see theirs)
      tags: &ref_9
        - jobs
  /shanda/jobs/{id}:
    get:
      operationId: JobsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponseDto'
        '403':
          description: Forbidden.
        '404':
          description: Job not found.
      security: *ref_8
      summary: Get a specific job by ID
      tags: *ref_9
  /shanda/jobs/{id}/status:
    put:
      operationId: JobsController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobStatusDto'
      responses:
        '200':
          description: Job status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponseDto'
        '403':
          description: Forbidden.
        '404':
          description: Job not found.
      security: *ref_8
      summary: Update job status (providers only)
      tags: *ref_9
...
  /shanda/vehicles/{vehicleId}/maintenance:
    get:
      operationId: MaintenanceController_findAll
      parameters:
        - name: vehicleId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: List of maintenance records
      security: &ref_12
        - bearer: []
      summary: Get maintenance history for a vehicle
      tags: &ref_13
        - maintenance
    post:
      operationId: MaintenanceController_create
      parameters:
        - name: vehicleId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaintenanceRecordDto'
      responses:
        '201':
          description: Maintenance record created successfully
      security: *ref_12
      summary: Add a maintenance record to a vehicle
      tags: *ref_13
  /shanda/maintenance/{id}:
    delete:
      operationId: MaintenanceController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Maintenance record deleted successfully
      security: *ref_12
      summary: Delete a maintenance record
      tags: *ref_13
  /shanda/messages/conversations:
    get:
      operationId: MessagesController_getConversations
      parameters: []
      responses:
        '200':
          description: ''
      security: &ref_14
        - bearer: []
      summary: Get all conversations for the authenticated user
      tags: &ref_15
        - messages
  /shanda/messages/conversations/{jobId}:
    get:
      operationId: MessagesController_getConversation
      parameters:
        - name: jobId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security: *ref_14
      summary: Get messages for a specific job conversation
      tags: *ref_15
  /shanda/messages:
    post:
      operationId: MessagesController_sendMessage
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageDto'
      responses:
        '201':
          description: ''
      security: *ref_14
      summary: Send a message in a conversation
      tags: *ref_15
  /shanda/messages/conversations/{conversationId}/read:
    post:
      operationId: MessagesController_markAsRead
      parameters:
        - name: conversationId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      security: *ref_14
      summary: Mark all messages in a conversation as read
      tags: *ref_15
...
info:
  title: Shanda Automotive API
  description: Marketplace connecting vehicle owners with service providers
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:4201
    description: Local Development
  - url: https://api.shanda.com
    description: Production
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    UpdateProfileDto:
      type: object
      properties:
        name:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        businessName:
          type: string
        serviceTypes:
          type: array
          items:
            type: string
        yearsInBusiness:
          type: number
        shopAddress:
          type: string
        shopCity:
          type: string
        shopState:
          type: string
        shopZipCode:
          type: string
        serviceArea:
          type: array
          items:
            type: string
        isMobileService:
          type: boolean
        isShopService:
          type: boolean
        hourlyRate:
          type: number
        website:
          type: string
        onboardingComplete:
          type: boolean
...
    JobResponseDto:
      type: object
      properties:
        id:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174000
        requestId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174001
        ownerId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174002
        ownerName:
          type: string
          example: John Doe
        providerId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174003
        providerName:
          type: string
          example: Jane Smith Auto Service
        quoteId:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174004
        agreedPrice:
          type: number
          example: 150
        description:
          type: string
          example: Oil change and filter replacement
        status:
          enum:
            - pending
            - pending_confirmation
            - in_progress
            - completed
            - cancelled
          type: string
          example: pending
        scheduledDate:
          format: date-time
          type: string
          nullable: true
        completedAt:
          format: date-time
          type: string
          nullable: true
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - requestId
        - ownerId
        - ownerName
        - providerId
        - providerName
        - quoteId
        - agreedPrice
        - description
        - status
        - createdAt
        - updatedAt
    UpdateJobStatusDto:
      type: object
      properties:
        status:
          enum:
            - pending
            - pending_confirmation
            - in_progress
            - completed
            - cancelled
          type: string
          description: The new status of the job
      required:
        - status
...
    CreateMaintenanceRecordDto:
      type: object
      properties:
        serviceType:
          type: string
          description: Type of service performed
          example: Oil Change
        serviceDate:
          type: string
          description: Date when service was performed
          example: '2026-01-04'
        mileage:
          type: number
          minimum: 1
          description: Mileage at time of service (km)
          example: 50000
        cost:
          type: number
          description: Cost of service
          example: 89.99
        notes:
          type: string
          description: Additional notes about the service
        performedBy:
          type: string
          description: Name of shop or mechanic who performed service
      required:
        - serviceType
        - serviceDate
    CreateMessageDto:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
          description: Job ID for the conversation
          example: 123e4567-e89b-12d3-a456-426614174000
        content:
          type: string
          description: Message content
          example: Hi, what time works best for you tomorrow?
      required:
        - jobId
        - content
...